Trying out new style BuiltinPrims, starting with tuple
The prims for tuples now take an extra argument that denotes the class
to use for the produced value.  tuple.py passes in this class for use
within the body of the primitives.

This required changing core-syntax classes to have a VPointer instead of an
Address as their class position, with corresponding changes to get-class and
new-object in interp and util, respectively.