Egregious scoping bug.  We need to fix Addresses BADLY
Parable:  Don't use the same construct for two things.  The floaty nature of
Addresses is allowing us to do truly bizarre things to the fabric of objects.

I suspect this is why certain tests are hanging; None is getting set to a
non-falsy value in seq_iter, which causes various is? tests on None to fail,
which makes comparisons that wrap seq_iter never complete.